Custom number of shots and the fix to the path to nwq_qasm executable
In the original code, custom options in run() was bugged due to reversed parameters in hasattr() and custom number of shots was not working (always be 1024 despite how result dict shows) because the shot number was never in the cmd command.

Also, the package could not find the nwq_qasm executable previously. The path is also fixed.

ghz.py is modified to illustrate the custom options avalible.